fix: resolve critical bugs and refactor across all services
- account-service: swap swapped DB credentials in application.yml,
  fix wrong pagination metadata in findAll(), fix cache evict key
  in banAccount(), extract applyBan() to remove BanInfo duplication
- notification-service: add @EnableScheduling, fix NotificationScheduler
  logic (DLT publish + async batch via CompletableFuture.allOf),
  fix @EAN annotation on repository, add NotificationProducer with
  correct topic, fix AccountEventListener wrong method for deleted accounts
- order-service: add eventId to OrderCreatedEvent in OutboxMapper,
  fix whitespace typo in HTTP_CLIENT_PRODUCT_SERVICE_URL env var,
  refactor OrderSagaListener handlers into shared processHandledEvent helper,
  fix PENDING->CONFIRMED status on item-reserved event
- product-service: fix ProductMapper toOrderBatchResponse mappings,
  remove Pageable from findAllByIdWithLock, fix Swagger schema on batch endpoint
- shared-events: remove broken @Value in OutboxRelay, remove unused Logger imports

+++ account-service/src/main/java/dn/accountservice/dto/account/ListAccountResponse.java
@@ -15,8 +15,8 @@ public class ListAccountResponse {
     private List<AccountResponse> accounts;
 
     @Schema(description = "Количество страниц с аккаунтами")
-    private Integer totalPages;
+    private int totalPages;
 
     @Schema(description = "Количество элементов на одной странице")
-    private Integer totalElements;
+    private long totalElements;
 }
